Create Columbus Blue Jackets Project
Company A wants to take their employees out for a team building event and the event they chose was to attend a Columbus Blue Jackets game as a company. Company A wants to maximize the experience for everyone, and that includes wanting to see the Blue Jackets win the game. There are no restrictions on which day of the week to go, or which team to see the Blue Jackets play, only that Company A wants the best chances at seeing a win for their employees.